Part 1 without tests

+++ src/main/java/factory/PointFactory.java
@@ -4,5 +4,5 @@ import entity.Point;
 import entity.Pyramid;
 
 public interface PointFactory {
-    Point create(int x, int y, int z);
+    Point create(double x, double y, double z);
 }
