+++ src/main/java/factory/impl/PointFactoryImpl.java
@@ -5,7 +5,7 @@ import entity.Pyramid;
 import factory.PointFactory;
 
 public class PointFactoryImpl implements PointFactory {
-    public Point create(int x, int y, int z) {
+    public Point create(double x, double y, double z) {
         return new Point(x, y, z);
     }
 }
